fix(system-agent): finish passive QR host integration
